+++ addons/yat/src/scenes/overlay/components/terminal/components/input/Input.cs
@@ -23,17 +23,32 @@ namespace YAT.Scenes.Overlay.Components.Terminal
 		/// <param name="command">The command to be submitted.</param>
 		private void OnTextSubmitted(string command)
 		{
+			// If the input string starts with a specified character,
+			// treat it as a method call on the selected node
+			if (command.StartsWith('$'))
+			{
+				_yat.Terminal.SelectedNode.ParseAndCallMethods(command[1..]);
+				AddToTheHistory(command);
+				Clear();
+				return;
+			}
+
 			var input = TextHelper.SanitizeText(command);
 			input = TextHelper.ConcatenateSentence(input);
 
 			if (input.Length == 0 || _commandManager.Locked) return;
 
-			_yat.HistoryNode = null;
-			_yat.History.AddLast(command);
-			if (_yat.History.Count > _yat.Options.HistoryLimit) _yat.History.RemoveFirst();
+			AddToTheHistory(command);
 
 			_commandManager.Run(input);
 			Clear();
+		}
+
+		private void AddToTheHistory(string command)
+		{
+			_yat.HistoryNode = null;
+			_yat.History.AddLast(command);
+			if (_yat.History.Count > _yat.Options.HistoryLimit) _yat.History.RemoveFirst();
 		}
 
 		/// <summary>